==> scripts/Kim-et-al-2023/Sandhill_Crane-Kim_etal_2023/0.calculate.UDs.Rmd <==
---
title: "0.calculate.UDs"
author: "Dennis Kim"
date: '2022-04-06'
output: html_document
---

```{r setup, include=FALSE}
# knit setting
library(knitr)
knitr::opts_chunk$set(echo = TRUE)
#options(width=150)
#opts_chunk$set(fig.width=18,fig.height=10)
```

# Preamble 

call libraries
```{r packages, message =FALSE, warning = FALSE}
# data calling & wrangling 
library(here)
library(readr)
library(tidyr)
library(purrr)
library(dplyr)
library(lubridate)

# spatial analysis 
library(raster)
library(sp)
library(sf)
library(rgdal)

# movement analysis 
library(amt)
library(ctmm)

# visualization
library(ggplot2)
```

# Read each year filtered summer crane postional data 

import the filtered summer crane data and select one individual "5A"
```{r import per yr data}
# call summer data 
crane <- read_rds(here("data", "crane_tot.Rdata"))

# add a column that has julian date information 
crane$JD <- yday(crane$timestamp)

# subset the data for only 1 individual - 5A
crane.5a <- crane %>% dplyr::filter(id == "5A")

# look at the 5A data structure 
head(crane.5a)
```

subset the data by year and convert them to telemetry object
```{r subset per yr}
# let's filter out the summer range (07/15 ~ 09/22)

# 2015
crane15 <- crane.5a %>% filter(year == "2015") %>% filter(timestamp > "2015-07-15 00:00:00" & timestamp < "2015-09-22 23:59:59") %>% as.telemetry()

# 2016
crane16 <- crane.5a %>% filter(year == "2016") %>% filter(timestamp > "2016-07-15 00:00:00" & timestamp < "2016-09-22 23:59:59") %>% as.telemetry()

# 2017
crane17 <- crane.5a %>% filter(year == "2017") %>% filter(timestamp > "2017-07-15 00:00:00" & timestamp < "2017-09-22 23:59:59") %>% as.telemetry()

# 2018
crane18 <- crane.5a %>% filter(year == "2018") %>% filter(timestamp > "2018-07-15 00:00:00" & timestamp < "2018-09-22 23:59:59") %>% as.telemetry()

# 2019
crane19 <- crane.5a %>% filter(year == "2019") %>% filter(timestamp > "2019-07-15 00:00:00" & timestamp < "2019-09-22 23:59:59") %>% as.telemetry()

# 2020 
crane20 <- crane.5a %>% filter(year == "2020") %>% filter(timestamp > "2020-07-15 00:00:00" & timestamp < "2020-09-22 23:59:59") %>% as.telemetry()
```

# calculate UDs pet yr 
```{r calculate UDs}
# facilitate good starting guesses for ctmm.fit (start with a proptotype hypothesis)
guess15 <- ctmm.guess(crane15, interactive = FALSE) 
guess16 <- ctmm.guess(crane16, interactive = FALSE) 
guess17 <- ctmm.guess(crane17, interactive = FALSE) 
guess18 <- ctmm.guess(crane18, interactive = FALSE) 
guess19 <- ctmm.guess(crane19, interactive = FALSE) 
guess20 <- ctmm.guess(crane20, interactive = FALSE) 

# fit continuous-time movement models 
fit15 <- ctmm.fit(crane15, guess15)
fit16 <- ctmm.fit(crane16, guess16) 
fit17 <- ctmm.fit(crane17, guess17)
fit18 <- ctmm.fit(crane18, guess18)
fit19 <- ctmm.fit(crane19, guess19)
fit20 <- ctmm.fit(crane20, guess20)

summary(fit15)

# compute occurrence distribution
UD15 <- occurrence(crane15, fit15)
UD16 <- occurrence(crane16, fit16)
UD17 <- occurrence(crane17, fit17)
UD18 <- occurrence(crane18, fit18)
UD19 <- occurrence(crane19, fit19)
UD20 <- occurrence(crane20, fit20)

# plot occurrence UD
plot(UD15, col.level = NA)
plot(UD16, col.level = NA)
plot(UD17, col.level = NA)
plot(UD18, col.level = NA)
plot(UD19, col.level = NA)
plot(UD20, col.level = NA)
```

# Export ctmm data formats to common raster formate and save 
```{r save data}
# sf object
UD15.sf <- as.sf(UD15)
UD16.sf <- as.sf(UD16)
UD17.sf <- as.sf(UD17)
UD18.sf <- as.sf(UD18)
UD19.sf <- as.sf(UD19)
UD20.sf <- as.sf(UD20)

# raster object (rasterize the probability density function)
UD15.raster <- raster(UD15, DF = "PDF")
UD16.raster <- raster(UD16, DF = "PDF")
UD17.raster <- raster(UD17, DF = "PDF")
UD18.raster <- raster(UD18, DF = "PDF")
UD19.raster <- raster(UD19, DF = "PDF")
UD20.raster <- raster(UD20, DF = "PDF")
```

```{r visualize the raster layers}
sp::plot(UD15.raster)
sp::plot(UD16.raster)
sp::plot(UD17.raster)
sp::plot(UD18.raster)
sp::plot(UD19.raster)
sp::plot(UD20.raster)
```


# save the sf and raster objects per year to shapefile
```{r save shapefile}
# sf
st_write(UD15.sf, "SummerCrane15_UD.shp")
st_write(UD16.sf, "SummerCrane16_UD.shp")
st_write(UD17.sf, "SummerCrane17_UD.shp")
st_write(UD18.sf, "SummerCrane18_UD.shp")
st_write(UD19.sf, "SummerCrane19_UD.shp")
st_write(UD20.sf, "SummerCrane20_UD.shp")

# raster
writeRaster(UD15.raster, filename = "UD15.tif")
writeRaster(UD16.raster, filename = "UD16.tif")
writeRaster(UD17.raster, filename = "UD17.tif")
writeRaster(UD18.raster, filename = "UD18.tif")
writeRaster(UD19.raster, filename = "UD19.tif")
writeRaster(UD20.raster, filename = "UD20.tif")
```

# Document Footer 
```{r session info}
sessionInfo()
```
